Feat: 엔트리 컴포넌트 분리
