+++ src/app/entry/[entryId]/components/position-entry.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useShallow } from 'zustand/react/shallow';
 import classNames from 'classnames';
